Test :  step3를 위한 인수테스트 생성

+++ src/main/java/nextstep/subway/favorite/ui/FavoriteController.java
@@ -5,6 +5,7 @@ import nextstep.subway.favorite.application.FavoriteService;
 import nextstep.subway.favorite.dto.FavoriteRequest;
 import nextstep.subway.favorite.dto.FavoriteResponse;
 import nextstep.subway.member.domain.LoginMember;
+import org.springframework.http.HttpStatus;
 import org.springframework.http.ResponseEntity;
 import org.springframework.web.bind.annotation.*;
 
@@ -24,6 +25,10 @@ public class FavoriteController {
     @PostMapping
     public ResponseEntity createFavorite(@RequestBody FavoriteRequest favoriteRequest, @AuthenticationPrincipal LoginMember loginMember) {
 
+        if (loginMember == null) {
+            return ResponseEntity.status(HttpStatus.UNAUTHORIZED).build();
+        }
+
         FavoriteResponse response = favoriteService.createFavorite(favoriteRequest, loginMember.getId());
         return ResponseEntity.created(URI.create("/favorites/"+response.getId())).body(response);
     }
@@ -31,12 +36,20 @@ public class FavoriteController {
     @GetMapping
     public ResponseEntity<List<FavoriteResponse>> inquiryFavorite(@AuthenticationPrincipal LoginMember loginMember) {
 
+        if (loginMember == null) {
+            return ResponseEntity.status(HttpStatus.UNAUTHORIZED).build();
+        }
+
         List<FavoriteResponse> responses = favoriteService.findByMemberId(loginMember.getId());
         return ResponseEntity.ok(responses);
     }
 
     @DeleteMapping("{id}")
     public ResponseEntity removeFavorite(@PathVariable Long id, @AuthenticationPrincipal LoginMember loginMember) {
+
+        if (loginMember == null) {
+            return ResponseEntity.status(HttpStatus.UNAUTHORIZED).build();
+        }
 
         favoriteService.removeFavorite(id, loginMember.getId());
         return ResponseEntity.noContent().build();
